+++ SkinTrade/app/src/main/java/com/example/skintrade/Model/Product.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 abstract class Product {
     abstract val id: Int
     abstract val name: String
-    abstract val price: Int
+    abstract val price: Double
     abstract val image: String
     abstract val description: String
 }
