+++ SkinTrade/app/src/main/java/com/example/skintrade/Model/Soundtrack.kt
@@ -3,13 +3,12 @@ package com.example.skintrade.Model
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
-
 @Serializable
 @SerialName("soundtrack")
 data class Soundtrack(
     override val id: Int,
     override val name: String,
-    override val price: Int,
+    override val price: Double,
     override val image: String,
     override val description: String,
     val author: String
